+++ docs/news.rst
@@ -25,8 +25,8 @@ Psycopg 3.3.0 (unreleased)
 
 - Add `pq.PGconn.used_gssapi` attribute and `Capabilities.has_used_gssapi()`
   function (:ticket:`#1138`).
-- Add support for the :pq:`PQfullProtocolVersion()` function, introduced in
-  libpq v18 (:ticket:`#1079`).
+- Add `ConnectionInfo.full_protocol_version` attribute,
+  `Capabilities.has_full_protocol_version()` function (:ticket:`#1079`).
 
 .. rubric:: Other changes